+++ drivers/dsp/dspbridge/mpu_driver/src/pmgr/io.c
@@ -61,7 +61,7 @@
 #include <iodefs.h>
 
 /*  ----------------------------------- Globals */
-static ULONG cRefs;
+static u32 cRefs;
 
 #if GT_TRACE
 static struct GT_Mask IO_DebugMask = { 0, 0 };	/* WCD IO Mask */
